Database Created
I have created the SQLServer database as the Unit 11 (Connected Acess)
suggests. It is added in the App_Data folder and its tables can be
accessed through the server explorer.

What's more I have created the Users table just to start with a
non-empty database, even the primary key and the data types are
specified in that table.

In addition, now the web application project contains a reference to the
class library one (both of them are in the same solution) so that we can
use the business entities and data access components in our app.