Project Profiles
Project Profiles are now loaded dynamically. The URL is being used for
passing the Project Title and also its Code so we can get it from the
DataBase. The Title, description, creator, creation data and state are
dynamic. The Credit label cannot be loaded yet because of some float/int
conflicts. The project comments system is not working neither.