Explore Projects improvement
The Explore Projects page shows the most recently created projects when
you first visit it. The most recents are the projects created on the
same day you entered the web. This is controlled using a variable stored
in the Global.asax file.

I have also fixed the results pages as they were not showing the correct
information when navigating through the results. This has been
controlled also using a variable created in the Global.asax file.

The Most Popular projects option in the left menu is not longer there.
Now we have a button over the results gridview which allows us to do
that search. The most popular are those projects whose total_bank is
equal or greater to the 0.9 of the maximun total_bank in the DDBB.