SignUp Fixed
Everyfield in the sign up is required right now. The email correctness
is also checked. I am trying to use a regular expression validator on
the name, lastname and username fields. Gender is the only field which
is not required right now but it will be checked as soon as I can do it.