Projects General update
Now when the user creates a project, a feedback text is shown.
The Explore Projects had a bug: while moving through the pages of the
Most Popular Projects  the results were showing incorrectly.
The ProjectDAC method Update has been change so it works properly now
with the state (should we keep this 'state' field?).
The Users Signup has been improved. Now, the name, last name and
username field formats are being checked.