Database Minor Changes
The new table Comments has been added to support the project comments
system.

The date fields of the Message table and Comments are now VARCHAR(20) to
fit a "dd/MM/yyyy hh:mm:ss" format datetime. We are testing it and we
will probably move all tables from the previous VARCHAR(11) "dd/MM/yyyy"
format to the new one.