Project Comments
First version of the project comments system. You can read comments in a
project profile. They show the user, the message and also the date. 10
comments per page. You can also write a comment. It will be shown as
soon as you postback the page (I am working on this). Added a couple of
lines to TODO list.