General Update
- ENTER was causing problems logging out the user when pressing it at
text box. I have added the Panel control and set the default button.
- SESSION was having some problems to identify users with cookies. I
have made a public static method at site.master.cs which loads the
cookie into the Session. Every private page calls that method.
- STYLE has been changed to textbox. They all use the same style now.
Added the groove border style :).
- '!' SYMBOL was not possible at some textboxes. Now, they are allowed
in some of them as project's messages and description.
- USERS EMAIL is not shown now. Every query has been updated so it gets
the user's nickname which is shown. This has been done in terms of
security and privacy.
- RESULTS shown at explore projects has been changed. They were no
showing the latest ones when the day has one digit. The DataBase stored
it using 0X but we were using X in the code here. Their style has also
changed.
- CONTRIBUTIONS is the most important update. Now, when you contribute
to the project three things happen: update the project (it has more
credits now), update the user (the user has less credits) and the
session variable, and also insert into Contributions the new
contributions.
Contributions is not working properly, it uses the date format
dd/MM/yyyy so users can contribute as many times as they want but the DB
does only accept one/day. It should be updated to include the time.