Statistical Singleton Class
We have added a Statistical Singleton class which uses the rest of the
DACs in the library to collect statistics about the website (number of
projects, number of contributions...).

What's more some methods have been added to the corresponding DACS and
the Master Page has been changed to use the singleton.

In addition, now the initial project creation contribution is inserted
as a normal contribution.