User Developments
I have implemented the development system in the webpage. The user can
develop any project. We are not handling with the GitHub right now but
we want to. Everytime you click on DEVELOP!, the developments entry will
be created.

At the user profile, this entry can be checked using My Developments.

There are some other changes related to regular expression validations.