General Comments
Added comments in some aspx pages and their cs files. I have also added
comments to the CommentBE and DAC in the most important methods which
are being used right now.
We are not using Global.asax anymore for storing a variable when
searching as it is share by all the users. Instead, we are using
Session.
I have also added some lines to the TODO file.